posterbator: support mapping of each selected element to palette
This does not sound useful! BUT! I use Lightburn for cutting and
it maps each layer with a color and there is no option to change
that.

In this patch I provide an option to map each selection element to
the lightburn color palette in order to have similar layers stack as
I see it in Inkscape and then in Lightburn after import.

The holes group is always colored with 20th color (the last one).

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>